Extract the Matcher factory to PathMatchers
