+++ test/java/org/opendatakit/briefcase/export/CsvFieldMappersTest.java
@@ -16,7 +16,6 @@
 
 package org.opendatakit.briefcase.export;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.stream.Collectors.toList;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
@@ -28,20 +27,16 @@ import static org.opendatakit.briefcase.export.Scenario.repeatGroup;
 import static org.opendatakit.briefcase.matchers.PathMatchers.exists;
 import static org.opendatakit.briefcase.reused.UncheckedFiles.list;
 
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Locale;
 import java.util.TimeZone;
 import java.util.UUID;
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
 import org.javarosa.core.model.DataType;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.opendatakit.briefcase.matchers.PathMatchers;
 import org.opendatakit.briefcase.reused.Pair;
 import org.opendatakit.briefcase.reused.UncheckedFiles;
 
@@ -278,33 +273,7 @@ public class CsvFieldMappersTest {
     Path outputAudit = scenario.getOutputDir().resolve(scenario.getFormName() + " - audit.csv");
 
     assertThat(outputAudit, exists());
-    assertThat(outputAudit, fileContains("line 1"));
-  }
-
-  private Matcher<Path> fileContains(String content) {
-    return new TypeSafeMatcher<Path>() {
-      private String actualContents;
-
-      @Override
-      protected boolean matchesSafely(Path item) {
-        try {
-          actualContents = new String(Files.readAllBytes(item), UTF_8);
-          return actualContents.contains(content);
-        } catch (IOException e) {
-          throw new RuntimeException(e);
-        }
-      }
-
-      @Override
-      public void describeTo(Description description) {
-        description.appendText("a file containing ").appendValue(content);
-      }
-
-      @Override
-      protected void describeMismatchSafely(Path item, Description mismatchDescription) {
-        mismatchDescription.appendText("was a file containing ").appendValue(actualContents);
-      }
-    };
+    assertThat(outputAudit, PathMatchers.fileContains("line 1"));
   }
 
   @Test
